+++ src/Data/AbstractAdapterRequest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Ifds\HttpAdapter\Data;
+namespace Ifds\Interchange\Data;
 
 use Illuminate\Support\Str;