+++ src/Events/ResponseReceived.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Ifds\HttpAdapter\Events;
+namespace Ifds\Interchange\Events;
 
 use Illuminate\Http\Client\Response;