+++ src/Exceptions/InterchangeException.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Ifds\Interchange\Exceptions;
+
+use RuntimeException;
+
+class InterchangeException extends RuntimeException {}